LayoutTests/fast/text-extraction/debug-text-extraction-interactions-expected.txt

PASS clickError is ""
PASS clickCount.textContent is "1"
PASS textInputError is ""
PASS textField.value is "bar"
PASS selectMenuItemError is ""
PASS select.value is "Three"
PASS selectTextError is ""
PASS getSelection().toString() is "Subject"
PASS scrollChildContainerError is ""
PASS childScroller.scrollTop is 100
PASS scrollPageError is ""
PASS document.scrollingElement.scrollTop is 100
PASS invalidActionError is "Failed to resolve nodeIdentifier 4294967293"
PASS clickDisabledError is "Click target is disabled"
PASS textInputReadonlyError is "Target is readonly"
PASS selectDisabledMenuItemError is "Select is disabled"
PASS readonlyField.value is "readonly"
PASS disabledSelect.value is "Beta"
PASS successfullyParsed is true

TEST COMPLETE
Test  Disabled
Subject

Hello world.

1

Foo bar
